I added more stuff to the carousel
The first picture is a placeholder. I'll replace it with a mock up of our site's stuff. Let me know what you guys think of the carousel? Does it work? I wanted the viewer the have some sort of an introduction to the site, for incase it's a first time user. I don't think all the buttons are necessary. I'll either change and spruce some up or I'll remove them completely. I still want to add a play button on all the titles, and change the spacing. And I'll also add the nav-bar. What else can a put on the page? Oh it's barely responsive at the moment.