Add to watchlist function works
The add to bloody watchlist thing works. I still have to make a template for the watchlist page's cards and then I have to add the info to the cards. I had so much trouble with this. the overlay was apparently causing some problems I didn't realize until far too late into the struggle. If it wasn't for Carlo and my sister TT... Oh yeah, the button is floating on top of the card now. Hehh I know... It's ugly but I needed a quick fix to just see if things are working. The overlay is really causing problems. IDK if people read this but if someone does, whyyyyyyyy would you choose to spend your life like this? TTTTTTTTTTTTTT Trauma.