Updates to the Homepage
I struggled a bit with the API stuff, but messed around with it until it worked. The fix was so easy. There was a typo or something in the API key. If I could add a facepalming emoji, I would. Then some cards where loading over the other cards. Again face palming time. But the biggest dumb-ass moment was when I couldn't figure out why the "most recent movies" cards didn't have posters... Dude, of course not. They don't have poster's yet because they're so new. The I had to battle with the footer. It was giving me problems, but it was an easy fix. I just added in 30vh height and boom. Coding truly is not my strong suite, but I tried.